requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.10.0
gunicorn==20.1.0
connexion==2.14.1
swagger-ui-bundle >= 0.0.2
python_dateutil >= 2.6.0
setuptools >= 21.0.0
Flask-SQLAlchemy==3.0.2
PyJWT==2.6.0